pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "magent-tui"
version = "0.1.0"
description = "基于 AutoGen + Textual 的多智能体协作 TUI 系统"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "m-agent" }]
dependencies = [
    "autogen-agentchat>=0.4.0",
    "autogen-ext[openai,anthropic]>=0.4.0",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "textual>=0.60",
    "rich>=13.7",
    "httpx>=0.27",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
]

[project.optional-dependencies]
litellm = ["autogen-ext[litellm]>=0.4.0", "litellm>=1.40"]
anthropic = ["anthropic>=0.34"]
claude-agent = ["claude-agent-sdk>=0.1.0"]
desktop-build = ["pyinstaller>=6.0"]

[project.scripts]
magent-tui = "magent_tui.main:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
